--- modules/consensus/info.go
@@ -1,72 +1,0 @@
-package consensus
-
-import (
-	"github.com/boltdb/bolt"
-
-	"github.com/NebulousLabs/Sia/types"
-)
-
-// ChildTarget returns the target for the child of a block.
-func (cs *ConsensusSet) ChildTarget(id types.BlockID) (target types.Target, exists bool) {
-	_ = cs.db.View(func(tx *bolt.Tx) error {
-		pb, err := getBlockMap(tx, id)
-		if err != nil {
-			return nil
-		}
-		target = pb.ChildTarget
-		exists = true
-		return nil
-	})
-	return target, exists
-}
-
-// EarliestChildTimestamp returns the earliest timestamp that the next block can
-// have in order for it to be considered valid.
-func (cs *ConsensusSet) EarliestChildTimestamp(id types.BlockID) (timestamp types.Timestamp, exists bool) {
-	// Error is not checked because it does not matter.
-	_ = cs.db.View(func(tx *bolt.Tx) error {
-		pb, err := getBlockMap(tx, id)
-		if err != nil {
-			return err
-		}
-		timestamp = earliestChildTimestamp(tx.Bucket(BlockMap), pb)
-		exists = true
-		return nil
-	})
-	return timestamp, exists
-}
-
-// GenesisBlock returns the genesis block.
-func (cs *ConsensusSet) GenesisBlock() types.Block {
-	return cs.blockRoot.Block
-}
-
-// InCurrentPath returns true if the block presented is in the current path,
-// false otherwise.
-func (cs *ConsensusSet) InCurrentPath(id types.BlockID) (inPath bool) {
-	_ = cs.db.View(func(tx *bolt.Tx) error {
-		pb, err := getBlockMap(tx, id)
-		if err != nil {
-			inPath = false
-			return nil
-		}
-		pathID, err := getPath(tx, pb.Height)
-		if err != nil {
-			inPath = false
-			return nil
-		}
-		inPath = pathID == id
-		return nil
-	})
-	return inPath
-}
-
-// StorageProofSegment returns the segment to be used in the storage proof for
-// a given file contract.
-func (cs *ConsensusSet) StorageProofSegment(fcid types.FileContractID) (index uint64, err error) {
-	_ = cs.db.View(func(tx *bolt.Tx) error {
-		index, err = storageProofSegment(tx, fcid)
-		return nil
-	})
-	return index, err
-}
