+++ modules/consensus/validtransaction_test.go
@@ -6,17 +6,73 @@ import (
 	"github.com/NebulousLabs/Sia/types"
 )
 
+// TestTryValidTransactionSet submits a valid transaction set to the
+// TryTransactionSet method.
+func TestTryValidTransactionSet(t *testing.T) {
+	if testing.Short() {
+		t.SkipNow()
+	}
+	cst, err := createConsensusSetTester("TestValidTransaction")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer cst.closeCst()
+	initialHash := cst.cs.dbConsensusChecksum()
+
+	// Try a valid transaction.
+	_, err = cst.wallet.SendSiacoins(types.NewCurrency64(1), types.UnlockHash{})
+	if err != nil {
+		t.Fatal(err)
+	}
+	txns := cst.tpool.TransactionList()
+	cc, err := cst.cs.TryTransactionSet(txns)
+	if err != nil {
+		t.Error(err)
+	}
+	if cst.cs.dbConsensusChecksum() != initialHash {
+		t.Error("TryTransactionSet did not resotre order")
+	}
+	if len(cc.SiacoinOutputDiffs) == 0 {
+		t.Error("consensus change is missing diffs after verifying a transction clump")
+	}
+}
+
+// TestTryInvalidTransactionSet submits an invalid transaction set to the
+// TryTransaction method.
+func TestTryInvalidTransactionSet(t *testing.T) {
+	if testing.Short() {
+		t.SkipNow()
+	}
+	cst, err := createConsensusSetTester("TestValidTransaction")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer cst.closeCst()
+	initialHash := cst.cs.dbConsensusChecksum()
+
+	// Try a valid transaction followed by an invalid transaction.
+	_, err = cst.wallet.SendSiacoins(types.NewCurrency64(1), types.UnlockHash{})
+	if err != nil {
+		t.Fatal(err)
+	}
+	txns := cst.tpool.TransactionList()
+	txn := types.Transaction{
+		SiacoinInputs: []types.SiacoinInput{{}},
+	}
+	txns = append(txns, txn)
+	cc, err := cst.cs.TryTransactionSet(txns)
+	if err == nil {
+		t.Error("bad transaction survived filter")
+	}
+	if cst.cs.dbConsensusChecksum() != initialHash {
+		t.Error("TryTransactionSet did not restore order")
+	}
+	if len(cc.SiacoinOutputDiffs) != 0 {
+		t.Error("consensus change was not empty despite an error being returned")
+	}
+}
+
 /*
-import (
-	"bytes"
-	"compress/gzip"
-	"crypto/rand"
-	"testing"
-
-	"github.com/NebulousLabs/Sia/crypto"
-	"github.com/NebulousLabs/Sia/types"
-)
-
 // TestValidSiacoins probes the validSiacoins method of the consensus set.
 func TestValidSiacoins(t *testing.T) {
 	if testing.Short() {
@@ -584,69 +640,3 @@ func TestValidTransaction(t *testing.T) {
 	}
 }
 */
-
-// TestTryValidTransactionSet submits a valid transaction set to the
-// TryTransactionSet method.
-func TestTryValidTransactionSet(t *testing.T) {
-	if testing.Short() {
-		t.SkipNow()
-	}
-	cst, err := createConsensusSetTester("TestValidTransaction")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cst.closeCst()
-	initialHash := cst.cs.dbConsensusChecksum()
-
-	// Try a valid transaction.
-	_, err = cst.wallet.SendSiacoins(types.NewCurrency64(1), types.UnlockHash{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	txns := cst.tpool.TransactionList()
-	cc, err := cst.cs.TryTransactionSet(txns)
-	if err != nil {
-		t.Error(err)
-	}
-	if cst.cs.dbConsensusChecksum() != initialHash {
-		t.Error("TryTransactionSet did not resotre order")
-	}
-	if len(cc.SiacoinOutputDiffs) == 0 {
-		t.Error("consensus change is missing diffs after verifying a transction clump")
-	}
-}
-
-// TestTryInvalidTransactionSet submits an invalid transaction set to the
-// TryTransaction method.
-func TestTryInvalidTransactionSet(t *testing.T) {
-	if testing.Short() {
-		t.SkipNow()
-	}
-	cst, err := createConsensusSetTester("TestValidTransaction")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cst.closeCst()
-	initialHash := cst.cs.dbConsensusChecksum()
-
-	// Try a valid transaction followed by an invalid transaction.
-	_, err = cst.wallet.SendSiacoins(types.NewCurrency64(1), types.UnlockHash{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	txns := cst.tpool.TransactionList()
-	txn := types.Transaction{
-		SiacoinInputs: []types.SiacoinInput{{}},
-	}
-	txns = append(txns, txn)
-	cc, err := cst.cs.TryTransactionSet(txns)
-	if err == nil {
-		t.Error("bad transaction survived filter")
-	}
-	if cst.cs.dbConsensusChecksum() != initialHash {
-		t.Error("TryTransactionSet did not restore order")
-	}
-	if len(cc.SiacoinOutputDiffs) != 0 {
-		t.Error("consensus change was not empty despite an error being returned")
-	}
-}
